updated readme

+++ launch_comfyUI_web_server.py
@@ -1,7 +1,7 @@
-from flask import Flask,send_file, render_template, request, Flask, send_from_directory, abort,jsonify
+from flask import Flask,send_file, render_template, request, Flask, after_this_request,jsonify
 from PIL import Image
 import os
-from workflow_api import style_workflow_api, upscale_workflow_api
+from workflow_api import style_workflow_api, upscale_workflow_api, config
 from utils import image_utils
 #from workflow_api import upscale_workflow_api
 import logging
@@ -9,11 +9,12 @@ from PIL import ImageEnhance, Image
 from flask_cors import CORS
 import tempfile
 import datetime
+import uuid
 
 
 # Define the paths for the input and output images
 parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
-UPSCALED_IMAGE_PATH = os.path.join(parent_dir,'comfy-web-server/static/output/')
+upscaled_image_path = os.path.join(parent_dir, config.upscaled_image_path)
 
 
 # Setup basic configuration for logging
@@ -36,57 +37,73 @@ def index():
     return render_template('index.html')
 
 
+
 @app.route('/process', methods=['POST'])
 def process():
     """Initial Image Processing Endpoint:
-     1. Convert to Oil Painting Style
-     2. Add custom color background 
-     3. Return the processed image url
-     4. Send the processed image to the frontend
-     """
+    1. Convert to Oil Painting Style
+    2. Add custom color background
+    3. Return the processed image url
+    4. Send the processed image to the frontend
+    """
 
     logging.info("Processing api reached...")
 
-    # check for image
+    # Check for image
     if 'image' not in request.files:
-        # Return an error to frontend if no image is uploaded
         return "No image uploaded", 400
     
     image = request.files['image']
 
-    # check for background color
-    background_color = request.form['background_color']
+    # Check for background color
+    background_color = request.form.get('background_color')
     if not background_color:
         return "No background color provided", 400
     
-    # Create a temporary file for the input image
-    with tempfile.NamedTemporaryFile(delete=False) as temp_input_file:
-        image.save(temp_input_file.name)
-        input_image_path = temp_input_file.name
+    # Create temporary files for input and processed images
+    input_image_path = None
+    processed_image_path = None
 
-    # Create a temporary file for the processed image
-    with tempfile.NamedTemporaryFile(delete=False) as temp_output_file:
-        processed_image_path = temp_output_file.name
-
-    # Save the image to the input image path
-    #image.save(os.path.abspath(INPUT_IMAGE_PATH))
     try:
+        with tempfile.NamedTemporaryFile(delete=False) as temp_input_file:
+            image.save(temp_input_file.name)
+            input_image_path = temp_input_file.name
+
+        with tempfile.NamedTemporaryFile(delete=False) as temp_output_file:
+            processed_image_path = temp_output_file.name
+
         logging.info("Processing the image...")
-        
+
         image_name = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
 
         # Process the image and save the processed image
-        processed_image_result_path  = style_workflow_api.process_image_with_comfy(
+        processed_image_result_path = style_workflow_api.process_image_with_comfy(
             input_image_path, processed_image_path, image_name, background_color
         )
         
         logging.info("Image processed...")
 
-        # Return a JSON response with the Base64-encoded image
-        return send_file(processed_image_result_path , mimetype='image/png')
+        # Return a JSON response with the image
+        return send_file(processed_image_result_path, mimetype='image/png')
+    
+    except Exception as e:
+        logging.error(f"Error processing image: {e}")
+        return {'error': str(e)}, 500
+
     finally:
-        os.remove(input_image_path)
-        os.remove(processed_image_path) 
+        # Close and remove the temporary files
+        if input_image_path:
+            try:
+                os.remove(input_image_path)
+            except Exception as e:
+                logging.error(f"Error removing input image file: {e}")
+
+        if processed_image_path:
+            try:
+                os.remove(processed_image_path)
+            except Exception as e:
+                logging.error(f"Error removing processed image file: {e}")
+
 
 # upscale image endpoint -> upscale the image, crop it using user defined values, and return the upscaled image
 @app.route('/upscale', methods=['POST'])
@@ -96,27 +113,28 @@ def upscale():
     3. Adjust brightness and contrast
     4. Resize the image based on the selected format
     5. Return the upscaled image url"""
-    
+
     logging.info("Upscale endpoint reached...")
 
     # Define a list of form fields
-    form_fields = ['brightness', 'contrast', 'format', 'crop_width', 'crop_height', 'crop_x', 'crop_y']
+    form_fields = ['brightness', 'contrast',  'crop_width', 'crop_height', 'crop_x', 'crop_y']
     # Use a dictionary comprehension to get the form data
     form_data = {field: request.form[field] for field in form_fields}
 
-    # check for image
+    # Check for image
     file = request.files['image']
     if file.filename == '':
         return "No image selected", 400
 
     logging.info("image received")
-    
+
     # Open the received image file
     input_image = Image.open(file.stream)
-    
+    logging.info("Image dimensions: %s", input_image.size)
+
     # Apply cropping
-    crop_x, crop_y = int(form_data['crop_x']), int(form_data['crop_y'])
-    crop_width, crop_height = int(form_data['crop_width']), int(form_data['crop_height'])
+    crop_x, crop_y = form_data['crop_x'], form_data['crop_y']
+    crop_width, crop_height =form_data['crop_width'], form_data['crop_height']
     cropped_img = image_utils.crop_image(input_image, crop_x, crop_y, crop_width, crop_height)
 
     # Adjust image brightness and contrast based on form_data
@@ -125,34 +143,41 @@ def upscale():
     adjusted_image = image_utils.adjust_brightness(cropped_img, brightness)
     adjusted_image = image_utils.adjust_contrast(adjusted_image, contrast)
 
-    # Create a temporary file for the adjusted image
-    with tempfile.NamedTemporaryFile(delete=False, suffix='.png') as temp_adjusted_file:
-        adjusted_image.save(temp_adjusted_file.name)
-        adjusted_image_path = temp_adjusted_file.name
-
-    # Create a temporary file for the upscaled image
-    with tempfile.NamedTemporaryFile(delete=False, suffix='.png') as temp_upscaled_file:
-        upscaled_image_path = temp_upscaled_file.name
+    # Create temporary files for the adjusted image
+    adjusted_image_path = None
 
     try:
-        image_name = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-        # upscale the image using process_image_with_comfy
+        with tempfile.NamedTemporaryFile(delete=False, suffix='.png') as temp_adjusted_file:
+            adjusted_image.save(temp_adjusted_file.name)
+            adjusted_image_path = temp_adjusted_file.name
+
+        image_id = str(uuid.uuid4())
+        image_name = datetime.datetime.now().strftime("%Y%m%d_%H%M%S") + image_id
+        # Upscale the image using process_image_with_comfy
         logging.info("Upscaling the image...")
 
         upscaled_image_result_path = upscale_workflow_api.process_image_with_comfy(
-            adjusted_image_path, os.path.abspath(UPSCALED_IMAGE_PATH), image_name, 3300
+            adjusted_image_path, os.path.abspath(upscaled_image_path), image_name, 3300
         )
         logging.info("Upscaling image complete!")
 
         # Construct the URL for the upscaled image
         image_url = f"output/{os.path.basename(upscaled_image_result_path)}"
-        
+
         # Return a JSON response with the image URL
         return jsonify({"image_url": image_url})
-    
+
+    except Exception as e:
+        logging.error(f"Error upscaling image: {e}")
+        return {'error': str(e)}, 500
     finally:
-        os.remove(adjusted_image_path)
-        os.remove(upscaled_image_path)
+        # Close and remove the temporary files
+        if adjusted_image_path:
+            try:
+                os.remove(adjusted_image_path)
+            except Exception as e:
+                logging.error(f"Error removing input image file: {e}")
+
 
 # If debug set to True the image processing will hang up before the image is processed
 # host must be set to '0.0.0.0' when hosting on a runpod server/dockerized container
